Move tricky workspace logic to a separate module
